design: #상담기록 장원태 [240219] - 상담일지 페이지 구성 및 스타일을 채팅방과 유사하게 변경
