+++ src/main/reactjs/src/components/chat/chattinglog/content/ChatLogContent.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ChatLogContent = () => {
+    return (
+        <div className='chatcontent fs_14 bor_red bg_red mt_10'>
+
+        </div>
+    );
+};
+
+export default ChatLogContent;
